refactor: create volunteer & organization folders
it should help to identify entities
