+++ src/main/java/com/voluntree/backend/domain/organization/Organization.java
@@ -1,4 +1,6 @@
-package com.voluntree.backend.domain;
+package com.voluntree.backend.domain.organization;
+
+import com.voluntree.backend.domain.User;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
