+++ src/main/java/com/voluntree/backend/domain/volunteer/Cpf.java
@@ -1,4 +1,4 @@
-package com.voluntree.backend.domain;
+package com.voluntree.backend.domain.volunteer;
 
 import java.util.List;
 import java.util.stream.Stream;
